CASMCMS-9225: Move PCS client to new paradigm
